Move pool into separate file from index
